orientation: Correctly identify orientation and doc type
Thanks to Lennart Karssen (who reached out via email) for making me
aware of the fact that xochitl (sanely) logs which files it opens.
Currently no mechanism is implemented to see whether the file actually
_is_ open (as opposed to just being the last entry in journalctl), but
that does not seem to be hard to add.

All that's left for proper orientation support is now to go through the
motions and correctly update all of the coordinates.

Related: https://github.com/isaacwisdom/RemarkableLamyEraser/issues/70#issuecomment-1712494730